web: Move PanicError enum to its own file
