+++ web/packages/core/src/internal/errors.tsx
@@ -1,0 +1,16 @@
+export enum PanicError {
+    Unknown,
+    CSPConflict,
+    FileProtocol,
+    InvalidWasm,
+    JavascriptConfiguration,
+    JavascriptConflict,
+    WasmCors,
+    WasmDownload,
+    WasmMimeType,
+    WasmNotFound,
+    WasmDisabledMicrosoftEdge,
+    InvalidSwf,
+    SwfFetchError,
+    SwfCors,
+}
